address review comments & small auth refactor
I'm creating a new function called `CurrentOrgName` which tries to fetch
the org name from the token and falls back to the `CurrentOrg` function
if not possible.

I also took the opportunity to remove some non-used dagger cache
initialization code
