+++ core/integration/cloud_test.go
@@ -19,15 +19,15 @@ func (CloudSuite) TestTraceURL(ctx context.Context, t *testctx.T) {
 	c := connect(ctx, t)
 
 	// depends on where the test runs - in an already nested test, we're *not* logged in
-	org, _ := auth.CurrentOrg()
+	org, _ := auth.CurrentOrgName()
 
 	url, err := c.Cloud().TraceURL(ctx)
-	if org == nil {
+	if org == "" {
 		requireErrOut(t, err, "no cloud organization configured")
 	} else {
 		require.NoError(t, err)
 		require.Contains(t, url, "https://dagger.cloud/")
-		require.Contains(t, url, org.Name)
+		require.Contains(t, url, org)
 	}
 }
 
@@ -35,7 +35,7 @@ func (CloudSuite) TestTraceURLNested(ctx context.Context, t *testctx.T) {
 	c := connect(ctx, t)
 
 	// depends on where the test runs - in an already nested test, we're *not* logged in
-	org, _ := auth.CurrentOrg()
+	org, _ := auth.CurrentOrgName()
 
 	src := `package main
 
@@ -51,11 +51,11 @@ func (m *Test) TraceURL(ctx context.Context) (string, error) {
 `
 	modGen := modInit(t, c, "go", src)
 	out, err := modGen.With(daggerCall("trace-url")).Stdout(ctx)
-	if org == nil {
+	if org == "" {
 		requireErrOut(t, err, "no cloud organization configured")
 	} else {
 		require.NoError(t, err)
 		require.Contains(t, out, "https://dagger.cloud/")
-		require.Contains(t, out, org.Name)
+		require.Contains(t, out, org)
 	}
 }
