+++ engine/telemetry/url.go
@@ -3,8 +3,6 @@ package telemetry
 import (
 	"context"
 	"fmt"
-	"os"
-	"strings"
 
 	"github.com/dagger/dagger/internal/cloud/auth"
 	"go.opentelemetry.io/otel/trace"
@@ -15,28 +13,9 @@ func URLForTrace(ctx context.Context) (url string, msg string, ok bool) {
 		return "", "", false
 	}
 
-	var orgName string
-	if cloudToken := os.Getenv("DAGGER_CLOUD_TOKEN"); cloudToken != "" {
-		// Try token auth first
-		token, ok := ParseDaggerToken(cloudToken)
-		if ok {
-			orgName = token.orgName
-		} else {
-			// DAGGER_CLOUD_TOKEN might be OIDC so try fetching
-			// the org from the state
-			org, err := auth.CurrentOrg()
-			if err != nil {
-				return "", "", false
-			}
-			orgName = org.Name
-		}
-	} else {
-		// Try OAuth next
-		org, err := auth.CurrentOrg()
-		if err != nil {
-			return "", "", false
-		}
-		orgName = org.Name
+	orgName, err := auth.CurrentOrgName()
+	if err != nil {
+		return "", "", false
 	}
 
 	if orgName == "" {
@@ -49,26 +28,4 @@ func URLForTrace(ctx context.Context) (url string, msg string, ok bool) {
 		trace.SpanContextFromContext(ctx).TraceID().String(),
 	)
 	return url, "", true
-}
-
-type daggerToken struct {
-	orgName string
-	token   string
-}
-
-func ParseDaggerToken(s string) (daggerToken, bool) {
-	s, ok := strings.CutPrefix(s, "dag_")
-	if !ok {
-		return daggerToken{}, false
-	}
-
-	orgName, token, ok := strings.Cut(s, "_")
-	if !ok {
-		return daggerToken{}, false
-	}
-
-	return daggerToken{
-		orgName: orgName,
-		token:   token,
-	}, true
 }
