+++ internal/cloud/auth/auth_test.go
@@ -1,4 +1,4 @@
-package telemetry
+package auth
 
 import (
 	"testing"
